+++ BitmovinYospacePlayer/Sources/BitmovinYospacePlayer/API/Events/BitmovinYospaceEvent.swift
@@ -1,0 +1,5 @@
+import BitmovinPlayerCore
+import Foundation
+
+@objc
+public protocol BitmovinYospaceEvent: NSObjectProtocol, Event {}