+++ BitmovinYospacePlayer/Sources/BitmovinYospacePlayer/API/Events/TrueXAdFreeEvent.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public class TrueXAdFreeEvent: NSObject, BitmovinYospaceEvent {
+    public var name = "onTrueXAdFree"
+    public var timestamp: TimeInterval = Date().timeIntervalSince1970
+}