+++ BitmovinYospacePlayer/Sources/BitmovinYospacePlayer/API/Events/YospaceEvents.swift
@@ -1,7 +1,7 @@
 import BitmovinPlayerCore
 import Foundation
 
-public class YospaceWarningEvent: NSObject, Event {
+public class YospaceWarningEvent: NSObject, BitmovinYospaceEvent {
     public var name: String
     public var message: String
     public var timestamp: TimeInterval
@@ -9,11 +9,11 @@ public class YospaceWarningEvent: NSObject, Event {
     public init(errorCode: YospaceErrorCode, message: String) {
         name = "Yospace Warning \(errorCode.rawValue.description)"
         self.message = message
-        timestamp = NSDate().timeIntervalSince1970
+        timestamp = Date().timeIntervalSince1970
     }
 }
 
-public class YospaceErrorEvent: NSObject, Event {
+public class YospaceErrorEvent: NSObject, BitmovinYospaceEvent {
     public var name: String
     public var message: String
     public var timestamp: TimeInterval
@@ -21,6 +21,6 @@ public class YospaceErrorEvent: NSObject, Event {
     public init(errorCode: YospaceErrorCode, message: String) {
         name = "Yospace Error \(errorCode.rawValue.description)"
         self.message = message
-        timestamp = NSDate().timeIntervalSince1970
+        timestamp = Date().timeIntervalSince1970
     }
 }