+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Swift Package Manager support
 - TruexAdRenderer support for tvOS (previously iOS only)
+- Support for using the Combine/Publisher based event listener approach
+  - `Player.events` will now properly include all Ad events
+- `Player.yospaceEvents` to listen to events from the Yospace integration (See `YospaceListener` for available events)
+- `BitmovinYospaceEvent` as parent type for all events from the Yospace integration
 
 ### Changed
 - **Migrated from CocoaPods to Swift Package Manager**
@@ -22,10 +26,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   // Now
   import BitmovinYospacePlayer
   ```
+- `YospaceListener.onTrueXAdFree` now gets a `TrueXAdFreeEvent` instance when called
+- Improved `YospaceListener` API
+  - Event callbacks are now optional allowing only implementing what's needed
+  - The sending `BitmovinYospacePlayer` instance is passed to each event callback
+- `YospaceListener.onTimelineChanged` got renamed to `YospaceListener.onAdTimelineChanged` to be consistent with the event name
 - Minimum Xcode version: 16.4
 - Minimum Swift version: 5.10
 - Project structure reorganized to follow SPM conventions (`Sources/BitmovinYospacePlayer/`)
-- `BitmovinPlayer` dependency to `3.100.0` and newer
+- `BitmovinPlayer` dependency to `3.103.0` and newer
 - `YOAdManagement-Release` dependency to `3.10.3`
 - `TruexAdRenderer-iOS` dependency to `3.5.1`
 - `TruexAdRenderer-tvOS` dependency to `3.15.2`